fix(ci): restore guardrail tooling checks
